clear up - dataset, train, test, visulzation
