+++ 2_train.sh
