+++ 4_frame_visualization.py
