--- gen_test.py
@@ -1,9 +1,0 @@
-import os
-from gen_dataset import *
-
-
-org_video_path = '/data/CAM_IO/'
-save_dir_path = '/data/CAM_IO/robot/images'
-
-
-gen_data(org_video_path, save_dir_path)
